Fixed integration test to override application beans
