+++ PetAdoptions/petsearch-java/src/main/resources/application.yml
@@ -4,16 +4,3 @@ spring:
 
 server:
   port: 80
-
-cloud:
-  aws:
-    region:
-      static: eu-central-1
-      auto: false
-    stack:
-      auto: false
-    credentials:
-      instance-profile: false
-#    credentials:
-#      accessKey: ABC
-#      secretKey: XYZ
