+++ PetAdoptions/petsearch-java/src/test/resources/application-test.yml
@@ -1,0 +1,16 @@
+spring:
+  profiles: test, default
+  main:
+    allow-bean-definition-overriding: true
+  autoconfigure:
+    exclude: org.springframework.cloud.aws.autoconfigure.context.ContextStackAutoConfiguration
+cloud:
+  aws:
+    region:
+      static: eu-central-1
+      auto: false
+    stack:
+      auto: false
+    credentials:
+      accessKey: ABC
+      secretKey: XYZ
